Fix schema and stop words template location
Buildout first initializes a recipe, then installs it and even later updates it.

As solr configuration depends on downloading solr
(done in a separate part with its own initialize, install and update),
checking for a file to exist on the other (download) part does not work on the initial phase,
but does on the install.

At the same time this recipe allows to override some templates,
so if the template is not overriden, an empty string is used (on the init phase).

Thus on the install phase, a correct path can be built.